feat(src): use I/O multiplex instead of subprocess
The single subprocess that handles HTTP requests is removed, instead
multiple HTTP server instances are created and handled via I/O
multiplexing. By doing so, binding the HTTP server to the specified
interfaces is enabled. The HTTP server sockets accept traffic only of
their address family. This makes the code to behave identical on Linux
and FreeBSD and removes the according known issues.