fix(src): set MULTICAST_IF option for send socket
IPv4 multicast traffic was sent on the default interface, not the one
which the socket was created for, even if you explicitly asked for that
interface.

Using bind(2) on Linux prevented this issue as well, but did not work on FreeBSD.
Although, the use of IP[V6]_MULTICAST_IF is not recommended, according to ip(4),
it works on FreeBSD and on Linux.

fixes #3