SOme usefdul changes to fix a small issue, and include basic file logging
Changes are self-evident from "preview". 

The FQDN issue is best shown by a copy of console (I don't know if this is something specific to FreeNAS or also an issue in FreeBSD generally, but let's fix it anyway):

```  # python3
Python 3.6.5 (default, Dec 20 2018, 21:27:40)
[GCC 4.2.1 Compatible FreeBSD Clang 6.0.0 (tags/RELEASE_600/final 326565)] on freebsd11
Type "help", "copyright", "credits" or "license" for more information.
>>> import socket
>>> print (socket.gethostname())
server2.localnet.lan
>>>```

Logging to file is needed if daemonised, which is the most likely use. Could be done better (log size limit, system log redirect) but at least it isn't just going to the console, and can be backgrounded neatly. logs don't seem to be huge, even if verbose, by nature of the protocol.