fix(src): only use host name part
According to user feedback, Python's gethostname() appears to return the
FQDN on FreeNAS. However, on Windows the name displayed in the 'Network'
view usually only includes the host name part. It is technically valid
to use the FQDN in the discovery process, but it does not match with the
behavior the user expects.

Thanks to stilez for poiting that out. See #5