fix(src): perform RFC-compliant Content-Type check
RFC 2387 specifies Content-Type as a list starting with an identifier, so
the previously assertion that `Content-Type == 'application/soap+xml'` is
in fact not spec compliant. The revised code retains the intent but
merely checks that the type starts with the suitable mimetype identifier,
it may be followed by any amount of extra information we don't care about.

Specifically this fixes compatibility with SOAP implementations that
always send the optional charset parameter and therefore caused
false positives with the previous check.

tldr:
`Content-Type: application/soap+xml; charset="utf-8"` no longer causes
400 errors


Co-authored-by: Harald Sitter <sitter@kde.org>