feat(src): observe network address changes
Instead of determining the network addresses in use at startup and
assume them to be static over the whole runtime, addresses and changes
to them are now tracked via the introduced NetworkAddressMonitor class.
Consequently, a network interface can be specified which does not have
useable addresses yet. If such one is assigned at a later point in time,
the desired service type (WSD client or host) is started. This has two
benefits. First, a change in networks (in case of a WiFi, e.g.) makes
the host to be announced automatically and no restart of the application
is required. Second, temporary/tentative addresses are not used for
binding and wsdd waits for those to become ready for use. This is
especially required for IPv6 tentative addresses present, e.g., at boot
time due to Duplication Address Detection (DAD) being in progress.

This commit implements the class NetworkAddressMonitor for Linux using
netlink sockets. The FreeBSD implementation is subject to future work.

A downside of the current implementation is that when a network is left,
no Bye message can be sent. According attempts result in "bad file
descriptor" error message on Linux. Thus, the host still appear in the
network view of windows hosts. However, it did not work before anyways.

The modified implementation also changes the behavior slightly. Previous
to this commit, the program terminated at startup when no available
address has been found. Now, the program continues assuming that a
useable address will appear on any (specified) interface. However,
specifying the -4 and -6 option still cause the program to terminate
since no addresses will be useable anyways.

see #25
see #18, #29, #1