feat(src): implement address change monitor for FreeBSD (on amd64)
The RouteSocketAddressMonitor handles changes of addresses on FreeBSD.
The offset used for iterating through if_msghdr's addresses might
architecture dependent due to alignment and options at compile time of
the OS. Thus, it may fail at some (exotic?) architectures.

The FreeBSD implementation also allows a convenient way to filter for
unusable interfaces that are either loopback or not multicast-capable.

An essential drawback is that tentative (IPv6) addresses are reported to
the application and no further notification is sent when the address
becomes ready for use. Thus, the refactored code has technically no
benefit over the old one. To solve that problem, the code would need to
poll for changes of the address via ioctl but that's inconvenient in the
current architecture.

fixes #25 (but only for Linux, actually)