fix(src): set PID in Netlink messages to 0 (fixes #50)
Since we only use a single socket, the kernel takes care of making it
unique on the whole system (across pid namespaces, i.e. containers) when
the socket is bind.

In addition, the nlmsg_pid field in the nlmsghdr message header is
opaque to the kernel and there is no reason to use the PID there as
well.